DXP-1310 Leave only qty field from stock data
