+++ src/catalog/Api/LoadQuantityInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace StreamX\ConnectorCatalog\Api;
 
-interface LoadInventoryInterface
+interface LoadQuantityInterface
 {
     public function execute(array $indexData, int $storeId): array;
 }
