--- src/catalog/Model/LoadInventory.php
@@ -1,25 +1,0 @@
-<?php
-
-
-declare(strict_types=1);
-
-namespace StreamX\ConnectorCatalog\Model;
-
-use StreamX\ConnectorCatalog\Api\LoadInventoryInterface;
-use StreamX\ConnectorCatalog\Model\ResourceModel\Product\Inventory as InventoryResource;
-
-class LoadInventory implements LoadInventoryInterface
-{
-    private InventoryResource $resource;
-
-    public function __construct(InventoryResource $resource)
-    {
-        $this->resource = $resource;
-    }
-
-    public function execute(array $indexData, int $storeId): array
-    {
-        $productIds = array_keys($indexData);
-        return $this->resource->loadInventory($productIds);
-    }
-}
